renamed from GaugeChart to Gauge
